// src/models/PayrollSinergy.ts
export interface PayrollSinergy {
  IdPayrollSinergy: number;
  namePayrollSinergy: string | null;
  codePayrollSinergy: string | null;
}

export interface PayrollSinergyCreate {
  namePayrollSinergy: string | null;
  codePayrollSinergy: string | null;
}

export interface PayrollSinergyUpdate {
  namePayrollSinergy: string | null;
  codePayrollSinergy: string | null;
}